requirements.txt
python-telegram-bot[job-queue]
sqlalchemy
python-dotenv
psycopg2-binary
